[C] set up review components
