+++ components/questions/Review/Widget/ColorFilterTool/styles.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import ColorFilterDisplay from "@/components/containers/ColorFilterDisplay";
+
+export const Display = styled(ColorFilterDisplay)`
+  margin-block-start: 1em;
+`;
