+++ components/questions/Review/index.ts
@@ -1,0 +1,4 @@
+export { default as Text } from "./Text";
+export { default as Select } from "./Select";
+export { default as Widget } from "./Widget";
+export { default as Inline } from "./Inline";
