--- components/questions/SimpleQuestion/Review/Text/styles.ts
@@ -1,17 +1,0 @@
-"use client";
-import styled from "styled-components";
-
-export const TextReview = styled.span`
-  background-color: var(--review-input-background-color);
-  border: 1px solid var(--review-input-border-color);
-  border-radius: 5px;
-  color: var(--review-input-color);
-  display: flex;
-  align-items: center;
-  font-size: 1rem;
-  margin-block-start: 1em;
-  padding: 0;
-  padding-inline: 1ch;
-  width: 100%;
-  min-height: 2rem;
-`;
