--- components/questions/SimpleQuestion/Review/Widget/styles.ts
@@ -1,8 +1,0 @@
-"use client";
-
-import styled from "styled-components";
-
-export const Container = styled.div`
-  display: block;
-  margin-block-start: 1em;
-`;
